requirements.txt
kafka-python
avro
